+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Env vars being set or unset for the app on deploy update errors
 - Typo on server pkg
 - [server] Return all deploy errors to the client
+- [client] teresaignore on Windows
 
 ## [0.11.0] - 2017-12-11
 ### Added